v0.3.0.3 外部短絡抵抗の全パターンを自動計算するよう修正
外部短絡抵抗の全パターンを自動計算するよう修正

+++ python/calc2/examples/short_test_webapp/apps/short_test_ini.py
@@ -103,7 +103,7 @@ def short_test_ini():
         if external_resistances == ['']:
             external_resistances = [0]
         else:
-            external_resistances = list(map(int, external_resistances))
+            external_resistances = list(map(float, external_resistances))
 
         st = ShortTest()
         # 一覧表を作成・保存
